requirements.txt
pymongo
aiohttp
quart
nbt
wrapt
python-dotenv
redis